1125 move checkbox helpers to helpers file
