+++ src/js/dp/search/_filter_handlers.js
@@ -1,35 +1,11 @@
 import { gtmDataLayerPush } from '../../utilities';
-import { switchSearchMarkup } from './_helpers';
+import { switchSearchMarkup, switchContentTypeFilterCheckbox, switchTopicFilterCheckbox } from './_helpers';
 import {
   clearValidation, validateDateFieldset, setFormValidation, validateDateRange,
 } from '../validation/validation';
 
 const searchContainer = document.querySelector('.search__container');
 const { title } = document;
-
-export const switchContentTypeFilterCheckbox = (paramsArray) => {
-  // get current param
-  const url = new URL(window.location.href);
-
-  // build new param
-  paramsArray.forEach((param) => {
-    if (!('isChecked' in param) || !('filterName' in param)) return;
-    if (param.isChecked) {
-      url.searchParams.append('filter', param.filterName);
-    } else {
-      const tmpValues = url.searchParams.getAll('filter').filter((e) => e !== param.filterName);
-      url.searchParams.delete('filter');
-      if (tmpValues.length !== 0) {
-        tmpValues.forEach((x) => {
-          url.searchParams.append('filter', x);
-        });
-      }
-    }
-  });
-
-  // make the change to the markup
-  switchSearchMarkup(searchContainer, url, true);
-};
 
 export const contentTypeCheckboxHandler = () => {
   [
@@ -56,7 +32,7 @@ export const contentTypeCheckboxHandler = () => {
           return checkbox;
         },
       );
-      switchContentTypeFilterCheckbox(paramsArray);
+      switchContentTypeFilterCheckbox(paramsArray, searchContainer);
 
       // Google Tag Manager
       gtmDataLayerPush({
@@ -69,7 +45,7 @@ export const contentTypeCheckboxHandler = () => {
       item.addEventListener('change', async (e) => {
         switchContentTypeFilterCheckbox([
           { isChecked: e.target.checked, filterName: e.target.value },
-        ]);
+        ], searchContainer);
         const filter = topFilter;
         filter.checked = theChildren.some((x) => x.checked);
 
@@ -82,38 +58,6 @@ export const contentTypeCheckboxHandler = () => {
       });
     });
   });
-};
-
-export const switchTopicFilterCheckbox = (paramsArray) => {
-  // get current param
-  const url = new URL(window.location.href);
-  paramsArray.forEach((param) => {
-    if (!('isChecked' in param) || !('topics' in param) || !('strParamType' in param)) return;
-    const { strParamType } = param;
-    const tmpValues = url.searchParams.getAll(strParamType);
-    url.searchParams.delete(strParamType);
-    if (tmpValues.length <= 1) {
-      if (param.isChecked) {
-        if (tmpValues.length === 0) {
-          tmpValues.push(param.topics);
-          url.searchParams.append(strParamType, tmpValues);
-        } else {
-          const tmpValue = tmpValues[0].split(',');
-          tmpValue.push(param.topics);
-          url.searchParams.append(strParamType, tmpValue);
-        }
-      } else if (tmpValues.length === 1) {
-        const tmpValue = tmpValues[0]?.split(',');
-        const tmpParam = tmpValue?.filter((e) => e !== param.topics);
-        if (tmpParam?.length !== 0) {
-          url.searchParams.append(strParamType, tmpParam);
-        }
-      }
-    }
-  });
-
-  // make the change to the markup
-  switchSearchMarkup(searchContainer, url, true);
 };
 
 export const topicFilterCheckboxHandler = () => {
@@ -142,7 +86,7 @@ export const topicFilterCheckboxHandler = () => {
           return checkbox;
         },
       );
-      switchTopicFilterCheckbox(paramsArray);
+      switchTopicFilterCheckbox(paramsArray, searchContainer);
 
       // Google Tag Manager
       gtmDataLayerPush({
@@ -155,7 +99,7 @@ export const topicFilterCheckboxHandler = () => {
       item.addEventListener('change', async (e) => {
         switchTopicFilterCheckbox([
           { isChecked: e.target.checked, topics: e.target.value, strParamType: 'topics' },
-        ]);
+        ], searchContainer);
         const filter = topicFilter;
         filter.checked = theChildren.some((x) => x.checked);
